@@ -1,9 +1,25 @@
-import "./App.css";
-import Map from "./Map";
-//import Layout from "./componets/layout/Layout";
+import logo from './logo.svg';
+import './App.css';
 
 function App() {
-  return <Map></Map>;
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
 }
 
 export default App;
